# contracts/openapi/chat/openapi_schema.json
{
    "info": {
        "title": "trip-chat-service",
        "version": "1.0.0"
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ],
    "paths": {
...
        "/api/v1/messages/{message_id}": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "GetMessage",
                "operationId": "ApiV1MessagesMessageIdGetMessage",
                "parameters": [
                    {
                        "name": "message_id",
                        "in": "path",
                        "schema": {
                            "type": "string",
                            "description": "Message ID."
                        },
                        "description": "Message ID.",
                        "required": true,
                        "deprecated": false
                    },
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "deprecated": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request fulfilled, document follows",
                        "headers": {},
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Message"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request syntax or unsupported method",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "extra": {
                                            "additionalProperties": {},
                                            "type": [
                                                "null",
                                                "object",
                                                "array"
                                            ]
                                        }
                                    },
                                    "type": "object",
                                    "required": [
                                        "detail",
                                        "status_code"
                                    ],
                                    "description": "Validation Exception",
                                    "examples": [
                                        {
                                            "status_code": 400,
                                            "detail": "Bad Request",
                                            "extra": {}
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
        "/api/v1/messages": {
            "get": {
                "tags": [
                    "Messages"
                ],
                "summary": "ListMessages",
                "operationId": "ApiV1MessagesListMessages",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "deprecated": false
                    },
                    {
                        "name": "conversation_id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "deprecated": false,
                        "allowEmptyValue": false,
                        "allowReserved": false
                    },
                    {
                        "name": "results_per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        },
                        "required": true,
                        "deprecated": false,
                        "allowEmptyValue": false,
                        "allowReserved": false
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "schema": {
                            "oneOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Base64-encoded Message ID."
                        },
                        "description": "Base64-encoded Message ID.",
                        "required": false,
                        "deprecated": false,
                        "allowEmptyValue": false,
                        "allowReserved": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request fulfilled, document follows",
                        "headers": {},
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CursorPagination_str_Message_"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request syntax or unsupported method",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "extra": {
                                            "additionalProperties": {},
                                            "type": [
                                                "null",
                                                "object",
                                                "array"
                                            ]
                                        }
                                    },
                                    "type": "object",
                                    "required": [
                                        "detail",
                                        "status_code"
                                    ],
                                    "description": "Validation Exception",
                                    "examples": [
                                        {
                                            "status_code": 400,
                                            "detail": "Bad Request",
                                            "extra": {}
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "deprecated": false
            }
        },
...
        "/api/v1/messages/stream": {
            "post": {
                "tags": [
                    "Messages"
                ],
                "summary": "StreamMessage",
                "operationId": "ApiV1MessagesStreamStreamMessage",
                "parameters": [
                    {
                        "name": "X-User-Id",
                        "in": "header",
                        "schema": {
                            "type": "string"
                        },
                        "required": true,
                        "deprecated": false
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SendMessageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Google ADK Event / Message",
                        "headers": {},
                        "content": {
                            "text/event-stream": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request syntax or unsupported method",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer"
                                        },
                                        "detail": {
                                            "type": "string"
                                        },
                                        "extra": {
                                            "additionalProperties": {},
                                            "type": [
                                                "null",
                                                "object",
                                                "array"
                                            ]
                                        }
                                    },
                                    "type": "object",
                                    "required": [
                                        "detail",
                                        "status_code"
                                    ],
                                    "description": "Validation Exception",
                                    "examples": [
                                        {
                                            "status_code": 400,
                                            "detail": "Bad Request",
                                            "extra": {}
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "deprecated": false
            }
        }
    },
    "components": {
        "schemas": {
            "Author": {
                "properties": {
                    "role": {
                        "type": "string",
                        "enum": [
                            "user",
                            "agent"
                        ]
                    },
                    "name": {
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Optional name of the user/agent."
                    }
                },
                "type": "object",
                "required": [
                    "role"
                ],
                "title": "Author",
                "description": "Author of the Message."
            },
...
            "CursorPagination_str_Message_": {
                "properties": {
                    "items": {
                        "items": {
                            "$ref": "#/components/schemas/Message"
                        },
                        "type": "array",
                        "description": "List of data being sent as part of the response."
                    },
                    "results_per_page": {
                        "type": "integer",
                        "description": "Maximal number of items to send."
                    },
                    "cursor": {
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Designating the last identifier in the given items."
                    }
                },
                "type": "object",
                "required": [
                    "items",
                    "results_per_page"
                ],
                "title": "CursorPagination[str, Message]"
            },
            "DataPart": {
                "properties": {
                    "data": {
                        "additionalProperties": {},
                        "type": "object",
                        "description": "Structured data content."
                    },
                    "kind": {
                        "type": "string",
                        "const": "data",
                        "default": "data"
                    },
                    "metadata": {
                        "oneOf": [
                            {
                                "additionalProperties": {},
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "data"
                ],
                "title": "DataPart"
            },
            "FilePart": {
                "properties": {
                    "file": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/FileWithUri"
                            },
                            {
                                "$ref": "#/components/schemas/FileWithBytes"
                            }
                        ]
                    },
                    "kind": {
                        "type": "string",
                        "const": "file",
                        "default": "file"
                    },
                    "metadata": {
                        "oneOf": [
                            {
                                "additionalProperties": {},
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "file"
                ],
                "title": "FilePart"
            },
            "FileWithBytes": {
                "properties": {
                    "bytes": {
                        "type": "string",
                        "description": "Base64-encoded file content."
                    },
                    "mime_type": {
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "examples": [
                            "application/pdf"
                        ]
                    },
                    "name": {
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "examples": [
                            "document.pdf"
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "bytes"
                ],
                "title": "FileWithBytes"
            },
            "FileWithUri": {
                "properties": {
                    "uri": {
                        "type": "string",
                        "description": "URI pointing to the file."
                    },
                    "mime_type": {
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "examples": [
                            "application/pdf"
                        ]
                    },
                    "name": {
                        "oneOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "examples": [
                            "document.pdf"
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "uri"
                ],
                "title": "FileWithUri"
            },
            "Message": {
                "properties": {
                    "message_id": {
                        "type": "string",
                        "description": "Unique identifier of the Message."
                    },
                    "conversation_id": {
                        "type": "string",
                        "description": "ID of the Conversation that this Message belongs to."
                    },
                    "author": {
                        "$ref": "#/components/schemas/Author"
                    },
                    "content": {
                        "items": {
                            "$ref": "#/components/schemas/Part"
                        },
                        "type": "array",
                        "description": "Content Parts, which are to be rendered, of the Message."
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "metadata": {
                        "oneOf": [
                            {
                                "additionalProperties": {},
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Optional key-value metadata. Useful for adding extra information."
                    }
                },
                "type": "object",
                "required": [
                    "author",
                    "conversation_id"
                ],
                "title": "Message"
            },
            "Part": {
                "oneOf": [
                    {
                        "$ref": "#/components/schemas/TextPart"
                    },
                    {
                        "$ref": "#/components/schemas/FilePart"
                    },
                    {
                        "$ref": "#/components/schemas/DataPart"
                    }
                ],
                "type": "object",
                "title": "Part"
            },
            "SendMessageRequest": {
                "properties": {
                    "conversation_id": {
                        "type": "string",
                        "description": "Conversation ID."
                    },
                    "content": {
                        "items": {
                            "$ref": "#/components/schemas/Part"
                        },
                        "type": "array",
                        "description": "Content of the query Message."
                    },
                    "metadata": {
                        "oneOf": [
                            {
                                "additionalProperties": {},
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "description": "Optional metadata for the query Message."
                    }
                },
                "type": "object",
                "required": [
                    "content",
                    "conversation_id"
                ],
                "title": "SendMessageRequest"
            },
...
            "TextPart": {
                "properties": {
                    "text": {
                        "type": "string",
                        "description": "String content of the part."
                    },
                    "kind": {
                        "type": "string",
                        "const": "text",
                        "default": "text"
                    },
                    "metadata": {
                        "oneOf": [
                            {
                                "additionalProperties": {},
                                "type": "object"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "text"
                ],
                "title": "TextPart"
            }
        }
    }
}